refactored visualizations.py. Added response models.
