+++ services/backend/src/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from src.routers import visualizations
+from src.routers.visualizations import router as visualizations_router
 from src.routers import auth as auth_router
 from src.chatbot import router as chatbot_router
 
@@ -22,7 +22,7 @@ app.add_middleware(
 )
 
 app.include_router(auth_router.router)
-app.include_router(visualizations.router)
+app.include_router(visualizations_router)
 app.include_router(chatbot_router.router)
 
 @app.get("/")
