+++ services/backend/src/routers/visualizations/models/knowledge.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+from .common import CorrelationRowBase
+
+
+class KnowledgeFunctionalityCorrelationRow(CorrelationRowBase):
+    knowledge_label: str
