+++ services/backend/src/routers/visualizations/models/perceptions.py
@@ -1,0 +1,9 @@
+from typing import List
+from pydantic import BaseModel, Field
+
+
+class TasksSupportDistributionResponse(BaseModel):
+    categories: List[str] = Field(default_factory=list)
+    values: List[int] = Field(default_factory=list)
+    domain: str
+    total: int = 0
