+++ services/backend/src/routers/visualizations/routes/sankey.py
@@ -1,0 +1,12 @@
+from fastapi import APIRouter
+
+from ..sankey_service import get_sankey_chart_data
+
+from ..models.sankey import SankeyDataResponse
+
+router = APIRouter()
+
+
+@router.get("/sankey-data", response_model=SankeyDataResponse)
+def sankey_data():
+    return get_sankey_chart_data()
